requirements.txt
langchain
langchain-community
langchain-openai
langchain-huggingface
chromadb
sentence-transformers
pypdf
python-dotenv
